황호연 _ 08 _ Enemy 베이스 로직 변경 , 게임/아이템 매니저생성
지상 적 경우에 로직을 다시 짜야함.
적이 아이템 드랍테이블을 가질 수 있음.
게임매니저와 아이템매니저가 씬에 들어감.
스크립트를 포함한 각종 파일들의 위치 변경

+++ Assets/Scripts/Item/ItemEnums.cs
@@ -5,15 +5,12 @@ using UnityEngine;
 /// <summary>
 /// 아이템 코드(ID)
 /// </summary>
-public enum ItemCode : byte
+public enum ItemCode : ushort
 {
     /// <summary>
     /// 소비 아이템 ( 즉발 )
     /// </summary>
     Apple = 0,
-
-
-
 
 
     /// <summary>
@@ -23,7 +20,7 @@ public enum ItemCode : byte
 
 
     /// <summary>
-    /// 판매 전용 아이템 
+    /// 판매 전용 아이템 ( 다이아몬드 )
     /// </summary>
     Diamond,
 
@@ -36,7 +33,7 @@ public enum ItemCode : byte
 
 
     /// <summary>
-    /// 악세서리 아이템 
+    /// 장신구 아이템 
     /// </summary>
 }
